Implement auth using Gate

+++ app/Http/Controllers/QuestionsController.php
@@ -63,6 +63,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question',$question)){
+            abort(403, 'Access Denied');
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -75,7 +78,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        //
+        if (\Gate::denies('update-question',$question)){
+            abort(403, 'Access Denied');
+        }
         $question->update($request->only('title','body'));
         return redirect('/questions')->with('success','Your question has been updated.');
     }
@@ -88,6 +93,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question',$question)){
+            abort(403, 'Access Denied');
+        }
         $question->delete();
         return redirect('/questions')->with('success','Your questions has been deleted.');
     }
